fix issues loading the recall responses of the participant
